# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "magic-manager",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/jobs/{name}": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Submit Job",
        "operationId": "submit_job",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "operationId": "get_job",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/events": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Job Events",
        "operationId": "job_events",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/event-stream": {
                "itemSchema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "string"
                    },
                    "event": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "integer",
                      "minimum": 0
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/edhrec/commanders": {
      "get": {
        "tags": [
          "edhrec"
        ],
        "summary": "Commanders",
        "operationId": "commanders",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommanderOption"
                  },
                  "title": "Response Commanders"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/edhrec/compare": {
      "get": {
        "tags": [
          "edhrec"
        ],
        "summary": "Compare",
        "operationId": "compare",
        "parameters": [
          {
            "name": "a",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "A"
            }
          },
          {
            "name": "b",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "B"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompareOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/collection": {
      "get": {
        "tags": [
          "collection"
        ],
        "summary": "Collection",
        "operationId": "collection",
        "parameters": [
          {
            "name": "families",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "minItems": 1,
              "title": "Families"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/collection/buy-list": {
      "post": {
        "tags": [
          "collection"
        ],
        "summary": "Collection Buy List",
        "operationId": "collection_buy_list",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyListIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuyListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BuyItem": {
        "properties": {
          "scryfall_id": {
            "type": "string",
            "title": "Scryfall Id"
          },
          "finish": {
            "type": "string",
            "enum": [
              "nonfoil",
              "foil"
            ],
            "title": "Finish"
          },
          "qty": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Qty",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "scryfall_id",
          "finish"
        ],
        "title": "BuyItem"
      },
      "BuyListIn": {
        "properties": {
          "target": {
            "type": "string",
            "enum": [
              "manapool",
              "tcgplayer",
              "moxfield",
              "plain"
            ],
            "title": "Target"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BuyItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "target",
          "items"
        ],
        "title": "BuyListIn"
      },
      "BuyListOut": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "lines": {
            "type": "integer",
            "title": "Lines"
          }
        },
        "type": "object",
        "required": [
          "text",
          "lines"
        ],
        "title": "BuyListOut"
      },
      "CollectionCardOut": {
        "properties": {
          "scryfall_id": {
            "type": "string",
            "title": "Scryfall Id"
          },
          "oracle_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oracle Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "family": {
            "type": "string",
            "title": "Family"
          },
          "set_code": {
            "type": "string",
            "title": "Set Code"
          },
          "collector_number": {
            "type": "string",
            "title": "Collector Number"
          },
          "rarity": {
            "type": "string",
            "title": "Rarity"
          },
          "type_line": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type Line"
          },
          "cmc": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cmc"
          },
          "color_identity": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Color Identity"
          },
          "released_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Released At"
          },
          "finishes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Finishes"
          },
          "owned": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Owned",
            "description": "finish -> owned copies"
          },
          "pledged": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Pledged",
            "description": "finish -> copies pledged to built decks"
          },
          "price_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Usd"
          },
          "price_usd_foil": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Usd Foil"
          },
          "image_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Uri"
          },
          "scryfall_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scryfall Url"
          },
          "treatment": {
            "type": "string",
            "title": "Treatment",
            "description": "Treatment codes (b, fa, shw, ext, sm, ff; '|'-joined); '' = plain"
          },
          "standard_frame": {
            "type": "boolean",
            "title": "Standard Frame"
          },
          "is_bulk": {
            "type": "boolean",
            "title": "Is Bulk"
          },
          "is_chase": {
            "type": "boolean",
            "title": "Is Chase"
          }
        },
        "type": "object",
        "required": [
          "scryfall_id",
          "oracle_id",
          "name",
          "family",
          "set_code",
          "collector_number",
          "rarity",
          "type_line",
          "cmc",
          "color_identity",
          "released_at",
          "finishes",
          "owned",
          "pledged",
          "price_usd",
          "price_usd_foil",
          "image_uri",
          "scryfall_url",
          "treatment",
          "standard_frame",
          "is_bulk",
          "is_chase"
        ],
        "title": "CollectionCardOut"
      },
      "CollectionOut": {
        "properties": {
          "families": {
            "items": {
              "$ref": "#/components/schemas/FamilySummaryOut"
            },
            "type": "array",
            "title": "Families"
          },
          "cards": {
            "items": {
              "$ref": "#/components/schemas/CollectionCardOut"
            },
            "type": "array",
            "title": "Cards"
          },
          "skipped": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skipped",
            "description": "Codes that didn't resolve to a family."
          }
        },
        "type": "object",
        "required": [
          "families",
          "cards"
        ],
        "title": "CollectionOut"
      },
      "CommanderOption": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "oracle_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oracle Id"
          },
          "type_line": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type Line"
          },
          "color_identity": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Color Identity"
          },
          "image_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Uri"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached",
            "description": "EDHREC commander page already in the local cache."
          }
        },
        "type": "object",
        "required": [
          "name",
          "oracle_id",
          "type_line",
          "color_identity",
          "image_uri",
          "cached"
        ],
        "title": "CommanderOption"
      },
      "CompareCardOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "oracle_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oracle Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "bucket": {
            "type": "string",
            "title": "Bucket"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "a_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "A Pct"
          },
          "b_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "B Pct"
          },
          "a_decks": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "A Decks"
          },
          "b_decks": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "B Decks"
          },
          "delta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta"
          },
          "synergy_a": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Synergy A"
          },
          "synergy_b": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Synergy B"
          },
          "trend_a": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trend A"
          },
          "trend_b": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trend B"
          },
          "type_line": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type Line"
          },
          "cmc": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cmc"
          },
          "mana_cost": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mana Cost"
          },
          "color_identity": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color Identity"
          },
          "rarity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rarity"
          },
          "lowest_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lowest Usd"
          },
          "lowest_usd_foil": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lowest Usd Foil"
          },
          "scryfall_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scryfall Id"
          },
          "image_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Uri"
          },
          "set_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Set Code"
          },
          "collector_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collector Number"
          },
          "scryfall_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scryfall Url"
          }
        },
        "type": "object",
        "required": [
          "name",
          "oracle_id",
          "slug",
          "bucket",
          "tags",
          "a_pct",
          "b_pct",
          "a_decks",
          "b_decks",
          "delta",
          "synergy_a",
          "synergy_b",
          "trend_a",
          "trend_b",
          "type_line",
          "cmc",
          "mana_cost",
          "color_identity",
          "rarity",
          "lowest_usd",
          "lowest_usd_foil",
          "scryfall_id",
          "image_uri",
          "set_code",
          "collector_number",
          "scryfall_url"
        ],
        "title": "CompareCardOut"
      },
      "CompareOut": {
        "properties": {
          "name_a": {
            "type": "string",
            "title": "Name A"
          },
          "name_b": {
            "type": "string",
            "title": "Name B"
          },
          "slug_a": {
            "type": "string",
            "title": "Slug A"
          },
          "slug_b": {
            "type": "string",
            "title": "Slug B"
          },
          "cards": {
            "items": {
              "$ref": "#/components/schemas/CompareCardOut"
            },
            "type": "array",
            "title": "Cards"
          }
        },
        "type": "object",
        "required": [
          "name_a",
          "name_b",
          "slug_a",
          "slug_b",
          "cards"
        ],
        "title": "CompareOut"
      },
...
      "FamilySummaryOut": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "printings": {
            "type": "integer",
            "title": "Printings"
          },
          "owned_printings": {
            "type": "integer",
            "title": "Owned Printings"
          },
          "owned_copies": {
            "type": "integer",
            "title": "Owned Copies"
          },
          "owned_usd": {
            "type": "number",
            "title": "Owned Usd"
          },
          "missing_printings": {
            "type": "integer",
            "title": "Missing Printings"
          },
          "missing_usd": {
            "type": "number",
            "title": "Missing Usd"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "printings",
          "owned_printings",
          "owned_copies",
          "owned_usd",
          "missing_printings",
          "missing_usd"
        ],
        "title": "FamilySummaryOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "inputs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Inputs"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "progress": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "artifacts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Artifacts"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "title",
          "inputs",
          "status",
          "created_at",
          "started_at",
          "finished_at",
          "progress",
          "summary",
          "artifacts",
          "error"
        ],
        "title": "JobOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}